fix android plugins nonce type

+++ plugins/nkn-sdk-flutter/android/src/main/kotlin/org/nkn/sdk/impl/Client.kt
@@ -318,7 +318,7 @@ class Client : IChannelHandler, MethodChannel.MethodCallHandler, EventChannel.St
         val duration = call.argument<Int>("duration")!!
         val meta = call.argument<String>("meta")
         val fee = call.argument<String>("fee") ?: "0"
-        val nonce = call.argument<Long>("nonce")
+        val nonce = call.argument<Int>("nonce")
 
         if (!clientMap.containsKey(_id)) {
             result.error("", "client is null", "")
@@ -328,7 +328,7 @@ class Client : IChannelHandler, MethodChannel.MethodCallHandler, EventChannel.St
         val transactionConfig = TransactionConfig()
         transactionConfig.fee = fee
         if (nonce != null) {
-            transactionConfig.nonce = nonce
+            transactionConfig.nonce = nonce.toLong()
         }
 
         viewModelScope.launch(Dispatchers.IO) {
@@ -350,7 +350,7 @@ class Client : IChannelHandler, MethodChannel.MethodCallHandler, EventChannel.St
         val identifier = call.argument<String>("identifier") ?: ""
         val topic = call.argument<String>("topic")!!
         val fee = call.argument<String>("fee") ?: "0"
-        val nonce = call.argument<Long>("nonce")
+        val nonce = call.argument<Int>("nonce")
 
         if (!clientMap.containsKey(_id)) {
             result.error("", "client is null", "")
@@ -360,7 +360,7 @@ class Client : IChannelHandler, MethodChannel.MethodCallHandler, EventChannel.St
         val transactionConfig = TransactionConfig()
         transactionConfig.fee = fee
         if (nonce != null) {
-            transactionConfig.nonce = nonce
+            transactionConfig.nonce = nonce.toLong()
         }
 
         viewModelScope.launch(Dispatchers.IO) {
